llm-inspector/backend/requirements.txt
# LLM Inspector v13.0 - Required Dependencies
# Managed by scripts/setup_dependencies.py — install via start.py or:
#   pip install -r requirements.txt

# === Core (always required) ===
numpy>=1.24.0
scipy>=1.10.0
cryptography>=41.0.0
PyYAML>=6.0               # SOURCES.yaml provenance registry

# === Recommended (used in main code paths, graceful fallbacks if absent) ===
requests>=2.28.0
scikit-learn>=1.3.0
SPARQLWrapper>=2.0.0      # Wikidata + DBpedia SPARQL (v13: promoted to recommended)
tiktoken>=0.7.0           # Tokenizer fingerprint probes

# === Optional (conditional imports — system degrades gracefully if absent) ===
# sentence-transformers>=2.2.0  # Local semantic similarity (needs torch)
# torch>=2.0.0                  # Required by sentence-transformers
# celery>=5.3.6                 # Distributed task queue
# redis>=5.0.1                  # Broker for Celery
# datasets>=2.19.0              # HuggingFace — suite_v13 benchmark fetching
# httpx>=0.27.0                 # Async HTTP for parallel KG queries
# fastapi>=0.104.0              # Production ASGI server (alternative)
# uvicorn>=0.24.0               # ASGI server
# jinja2>=3.1.0                 # Report HTML templates
# spacy>=3.7.0                  # Advanced NLP
# sqlalchemy>=2.0.0             # ORM (optional alternative to stdlib sqlite3)
